Update patch set 1
Patch Set 1:

This still leaves the scope of breaking a word, though. But such a scenario is unlikely unless there is a very long word which doesn't give any opportunity to browsers to break the line (since there is no white space). Such a long chain of characters without a white-space would likely be a link only in which case browsers will break the line which is acceptable for us.

Patch-set: 1
Reviewer: Gerrit User 1000897 <1000897@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0