Update patch set 3
Patch Set 3:

Other than that there is one issue that i have noticed, all the files requested by the admin page contain an extra '/' in its path.
Hence, compressed versions of those files are not being served (as of now).
Is that being kept there intentionally, or is that a bug?

Patch-set: 3