Update patch set 1
Patch Set 1: Code-Review+2 Verified+1

Not that excited to put a fix that is not BLOCKER/HIGHEST to the last RC; but as you pointed out, would be good to keep 2.1 and master in sync wrt. the trace file format, and the change is very local, so let's take it.

Patch-set: 1
Reviewer: Gerrit User 1000068 <1000068@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=+1
Label: Code-Review=+2