Update patch set 1
Patch Set 1:

The original design (as you might have found out) was in fact to include the host name. This was precisely to avoid a security hole (that I had anticipated when I put the host name in and added the following comment: "Keep the host as part of the key to close a potential security hole.").

Unfortunately after 2.0 release we discovered that this broke some cases where aliasing was used with the side-effect that saving overwrote the document back and forth, so it was decided to remove it.

(See bccu#2190 for more details.)

Patch-set: 1