Update patch set 1
Patch Set 1:

After thinking about this a bit more I think the better approach would be to use a unique host identifier (instead of the hostname in the URL, which could be aliased).

We could add it in the wopi getFileInfo response (and for non-wopi, or when failing that, we can fallback to using the hostname). So the file query would return some ID or name that is unique to the file storage that we can use to reference the file from then on.

I think with a simple host:port in the docKey (as in this patch) we are likely to break servers with aliases, which I expect is common.

Of course it's better to fix the security issue than worry about aliased access, but I'm trying to address both issues, which is tenable I believe.

Patch-set: 1
Reviewer: Gerrit User 1000871 <1000871@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0