Revert "don't hardcode lo_template_path, use the setting from loolwsd.xml"
This reverts commit ef23ab447d859aad7ac2824e7d61c2aec1cf42c8.
loolwsd.xml from Collabora Online 2.0 or older may contain wrong default for lo_template_path,
therefore after an upgrade, loolwsd would fail to start, or it would start wrong LibreOffice.
In Collabora's product this path is not really configurable, so it makes sense to hardcode it.
Upstream is more flexible in this regard.

Change-Id: I38f635f5fbff62fcd32aa1fd8f0b3c05b8030b08