Update patch set 2
Patch Set 2:

> It seems there are two or more commits on your local and you didn't squash them with same Change-Id before uploading this patch ?

So what had happened is that there was a patch that had changed the special handling of pdf, and this patch was reverted. Hence the effect you see.

But it certainly is removed from patch 2, and that's what counts.

Patch-set: 2
Reviewer: Gerrit User 1000871 <1000871@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0