Update patch set 2
Patch Set 2: Code-Review-1

(3 comments)

Great work ! I am happy with the approach. Made few quick comments, though I know this is WIP and you plan to fix them already :)

 > work in progress.
 > 
 > Current patch keeps track of changes(only) after the admin console
 > is opened, and instead of an icon for indication, can we just show
 > when the last modification was made and by which user(as in this
 > patch).

Nice idea - more the information we can show in admin-console, better. In addition to your proposal above to including when last modification was made and by which user, I think we would still need if the document is currently modified or not - this is necessary to give an idea to admin about state of the document people have left it in ( and kill it if necessary ;)

 > 
 > Even though i did no major edits, make check seems to be failing
 > because of some slide hash mismatch, due to some other commit
 > perhaps.
 > can anyone help me about how should i resolve that.

I'd ignore them, if they are not related to your patch. Sure, you can investigate if you want to but better focus on one thing at a time.

Patch-set: 2
Label: Code-Review=-1