Update patch set 3
Patch Set 3: Code-Review-1

(4 comments)

See inline comments.

And yes, this bug is about unsaved modification.

In the UI, since we will only be able to know if the document is modified or not and who saved the document last, a column titling 'Is Modified?' and the values like 'Yes' would be enough. To add the additional information of who saved the document last and at what time would be too much to fit in that column, I think :-) If we could get a hover-on feature that then displays the last saved by etc. information, that'd great but currently hover-on functionality is to tied to showing users only.

Patch-set: 3
Label: Code-Review=-1