Update patch set 4
Patch Set 4:

changes made

- removed username and time related to last modification
- now we only display a 'Yes' if a document is in modified state else a 'No'.

there seems to be an issue with saving documents though, when trying to save documents a pop up citing an 'unknown' error comes up. I went through the logs and there it seems that the response to uno save command is fine, can you help me with that :)

other than that, really sorry that it took me a day for this trivial task. Had the LOK header problem, and had to rebuild core, then had errors building core and ended up spending my whole day on this.

Patch-set: 4
Label: Verified=0