Update patch set 3
Patch Set 3:

Had quite some errors, which i initially overlooked.
tested it thoroughly and now it looks fine.

About the arrow, i tried it in chrome and it seems to be working fine.
once it happened that it wasn't working, i refreshed and it worked fine. wasn't able to reproduce it again.

Patch-set: 3
Label: Verified=0