Update patch set 1
Patch Set 1:

To confirm - why do we bother to pre-process the loleaflet.html ? this is something a customer can supply a pre-canned form of, and does.

It is only necessary to add <!-- remove this section to remove the about dialog --> in there.

From the Javascript - we can (surely?) detect that this is not present - and then not show the dialog.

So - I would expect this entire patch to be a single line of comment in the loleaflet.html, and perhaps another single line of Javascript to disable the relevant menu item.

Can you confirm that this does in fact disable the menu item ? perhaps that code is already there (?).

It's not really something we want to encourage peopel to do - so making it too easy is unfortunate I think.

Patch-set: 1
Reviewer: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0