Update patch set 3
Patch Set 3: Code-Review-1

> Had quite some errors, which i initially overlooked.
 > tested it thoroughly and now it looks fine.
 > 
 > About the arrow, i tried it in chrome and it seems to be working
 > fine.
 > once it happened that it wasn't working, i refreshed and it worked
 > fine. wasn't able to reproduce it again.

Hmm, no idea. Even this version's arrow doesn't work well on my chrome here on linux (Version 57.0.2987.133 (64-bit). But yeah it works nicely on firefox. Also, did you try on more vexing browsers like IE and MS Edge :-), may be the issue that I am seeing on chrome is also reproducible on IE and Edge ?

Is it possible to beautify the tabbed view there also, please ? Like appearance of a more tabbed view there instead of simple toggle buttons. See here[1] for how I have imagined it to be. 

And if this arrow thing is not easy-to-fix thing, I would be happy with a hover-on and see the list of documents too (with a heading so that user knows that the pop-up is a list of documents). Make sure that hover is set on the last column only, not on the full row.

Right now when in "user online" tab, I can right click and see a 'Kill' context-menu there. But we can't kill a session individually, lets remove this context menu when in this tab to avoid any confusion, we can file another bug for making it work separately.


[1] https://codepen.io/wizly/pen/BlKxo

Patch-set: 3