Update patch set 5
Patch Set 5:

so, i tried to change most things to bootstrap, the tabs and the list of users.
About the arrow thing, couldn't exactly find out the reason behind this, but many answers on stackoverflow suggested using a broader ancestor, and that seemed to work. Can you please check now.

Patch-set: 5