Update patch set 1
Patch Set 1:

I wonder - in most of the cases this patch removes lockLokDoc.unlock();, but in one occasion it stayed there...

What is the mechanics there?  It is unclear to me from the commit message what is going on exactly, and how is this commit fixing that? :-) - can you please explain it, ideally with an amended commit message?  Thank you!

Patch-set: 1
Label: Code-Review=0