Update patch set 1
Patch Set 1: Code-Review+2 Verified+1

Henry: Thank you for the patch!

In general, I would _strongly_ prefer to have this kind of unit tests directly in the core.git (ie. as native LOK unit tests), because ultimately what we are testing here is that the conflict info is sent, regardless of who is the user of LibreOfficeKit here.

Having said that, the more tests the merrier - so approving this :-)

Patch-set: 1
Reviewer: Gerrit User 1000068 <1000068@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=+1
Label: Code-Review=+2