Update patch set 2
Patch Set 2:

> I wonder - in most of the cases this patch removes
 > lockLokDoc.unlock();, but in one occasion it stayed there...
 > 
Updated the commit message. By removing the unlock call, we make sure that the call to notifyViewInfo is done under the lock. The one case that has it left, it's _after_ notifyViewInfo, because there is more processing happening there (the other cases are at the end of the function).

Patch-set: 2
Reviewer: Gerrit User 1000871 <1000871@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0