Update patch set 4
Patch Set 4: Code-Review-1

I am little worried about how the dialog obscures the document while you searching/replacing. Unlike libreoffice desktop, the dialog cannot be moved, it is fixed. Do you think it is possible somehow to integrate this replace/replaceall feature with the existing search box we have in the bottom toolbar, so that it doesn't obscure the document when user is doing this operation ?

Other than that, this patch seems to break the 'expansion' search feature:
* Have a document containing the text 'xxxx'
* Type 'x' in the search box
  * 'x' in the document gets selected
* Now type another x in the search box
  * The search result in the document should be expanded which was the case before this patch, but this patch breaks that behavior.

Patch-set: 4
Label: Code-Review=-1