Update patch set 8
Patch Set 8:

Ah, one small but important thing that must be done as a follow-up as soon as possible, unless there is a good reason.

I wonder why this is only enabled in writer, and not on calc and impress. Did you test it on calc and impress ? If it's working fine on impress and calc as well, please enter the menu entries there too. Please don't forget to test it there. I hope it works there fine as well.

Patch-set: 8