Update patch set 2
Patch Set 2:

> Actually I'd be very interested to know what happens when
 > 
 > * the document gets idle
 > * the server goes down
 > * the user clicks the "activate again" possibility? :-)
 > 
 > Either way, the code looks good to me.

Simple: when the document gets idle, it is unloaded and all clients disconnected, so when the server goes down, these clients (who were disconnected due to idleness) will not know about it.

Since the document idle notification now allows to reconnect on clicking (same as with inactive and server-recycling), when the user clicks a reconnection attempt will be made. But since the server is down the connection will fail and nothing will happen (from UI point of view). This is because such soft failures are expected by design and are ignored.

When the server is back online, the document will not be reloaded automatically, but once the user clicks in the doc, it will reload and allow the user to continue editing.

(I have verified the above on 2.1, so it's not just theoretical.)

Patch-set: 2
Reviewer: Gerrit User 1000871 <1000871@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0