Update patch set 5
Patch Set 5: Code-Review+2 Verified+1

(1 comment)

Looks good to me. I only managed to find a nitpick this time :-)

Can you also work in a separate commit on the beautification of these graphs as mentioned on the original bug report. Would be good to explore the d3 graph library. The documentation and examples are all over the internet, so that should help.

Patch-set: 5
Label: Verified=+1
Label: Code-Review=+2