Update patch set 2
Patch Set 2:

Hi,

I tried to find what the root cause for this was, but had no success.
For the time being, what we could do is disable search on every input change, so that the user can search only by clicking the 'next' and 'previous', just as it is in the desktop version.

Let me know if i should go ahead with it or not.

Patch-set: 2