Update patch set 6
Patch Set 6:

(1 comment)

I am confused regarding your use of hard-coded this._netAvgSize in the AdminSocketAnalytics.js file. Are we sure that value will be hard-coded to 10 irrespective of how far the data points in the data set are ? By default they are 5s, and size of the dataset is 100 points. How come we end up setting (100 - 10 + 1) size for the avg array. It'd be great if you could explain.

Patch-set: 6
Label: Verified=0