loleaflet: Allow users to repair at all times
Since during an edit conflict the undo/redo
buttons are dissabled (for the user who can't
undo/redo anymore) disabling the repair button
robs them of the ability to repair the document
to an earlier state (i.e. before the conflict).

This enables the repair button as soon as either
the undo or redo is enabled, whereupon it remains
enabled for the duration of the session.

Change-Id: Ibaf257155f721e2cef596e965d8812d381ab3a5b