Update patch set 1
Patch Set 1: Code-Review+2 Verified+1

Fair enough for 2.1.x - but - I'd really like us not to duplicate that thing shared_ptrs and all each time we want to broadcast something: that sucks =)

Instead we should tag sessions to be removed - and remove them in the tail end of the central 'poll' method while that is running - so we can be sure we're not iterating that array. ie. using the approach that the SocketPoll does itself. Then we can avoid the duplication etc. Any chance of doing that for master ? I'd like to reduce the complexity here.

Patch-set: 1
Reviewer: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=+1
Label: Code-Review=+2