Update patch set 4
Patch Set 4:

(9 comments)

Please see some comments inline.

Also, when I opened a document side by side in two browser windows, and started typing in one of them, it said that I am the editor now, even I didn't click on the 'Follow editor' button yet. Is that expected ? I had hoped that nothing would happen (i.e the situation would be same as it was before applying this patch) until I click on the 'Follow editor' button. But that doesn't seem to be the case.

Other than that, in one of the scenarios I encountered, it said that I am following someone else but when I made that some one else type, my view didn't jump to her cursor, so the userlist was lying to me :) This would be very confusing to the end user.

Patch-set: 4