Update patch set 6
Patch Set 6:

I am still not able to see the expected behavior in this patch. The view doesn't jump automatically to the place where editor is typing.

I have opened a document in two separate windows, one of them is the editor editing something in the first page. The other user is reading something in the fifth page or so. When the editor types something and this 'other' user is following it, the view of the 'other' user should jump to the cursor of the first user, which isn't happening.

Also, I don't like the idea of popping up the notification when a user becomes editor. That's quite distracting.

Anyhow, I think the design is still a bit baffling there to a normal user and we need something intuitive enough so that this feature, in an effort to make user's life more convenient, doesn't at least confuse them. Firstly, as mentioned above, lets do away with the notification of who become the editor. Just keep the editor information in the JS but don't tell the user until she clicks the user pop up. The plan is to keep this information only in the user list and that too in the symbolic form, not the written form whenever possible.

The design of the user list would be something like this. There would be an additional 'Always follow the editor (currently Aditya)' checkbox on the bottom of the user list, and it will do what it says - i.e follow the editor. Other than that, all of the user list items would be selectable and as soon as you select any of the user, two things will happen - 1) the 'Always follow the editor' checkbox will go unchecked and 2) the user list item will get selected which means we are explicitly following that user irrespective of whoever is the current editor. Also, this setting will prevail only until user clicks into the document and starts typing. As soon as the user starts typing, we will forget these settings and will stop following anyone, if we are already.

Also, when the user enters the document, there should be no following. It should work as it does right now in current master (without this patch). 

Of course before implementation, if there are any doubts feel free to speak up. Open to suggestions if you think any other design could make it better but not at the cost of simplicity - it shouldn't confuse users and it should be intuitive.

Patch-set: 6