Update patch set 6
Patch Set 6:

hi,
I am still working on this, small things are taking more time than anticipated.

I will be abandoning this patch.
In replacement i have uploaded 2 new patches, which solve small but apt issues.

- https://gerrit.libreoffice.org/#/c/39247/
this is only for following any user by clicking from the list.
There was a small issue, about not being able to differentiate between the cursor invalidation actually done by the user, or being done because of the placement of cursor(if A's cursor is ahead of B). Now it is solved and this works fine.

- https://gerrit.libreoffice.org/#/c/39299/
[WIP] overall it is working, but there are many small things that i am working on, should be complete by today. I have tried to change the design as discussed and it looks good to me, let me know if any changes need to be done in that regard.

Patch-set: 6