Update patch set 5
Patch Set 5:

(1 comment)

Just tried it, and it works mostly. Great work there ! However, I noticed some problems that we should fix.

Can we also hide the userlist popover when user clicks 'Follow editor'. The problem is that to hide that user will intuitively click into the document which will unfollow the editor. This would confuse the user. If the pop over is automatically hidden after the click, just like it happens when you click on a user, that would be great, I think.

And why does the name on top of the cursor keeps blinking as he types ? That is quite distracting, but a minor annoyance (not a high-priority to fix).

Also, in an automatic-follow mode in one of the views and after continuously typing only from some other view, the editor in the first view changed for me after 1.5 minutes !! That is huge time for a person to become an editor. I think some 10 seconds or so should be the time ?

Patch-set: 5