Update patch set 6
Patch Set 6:

> (1 comment)
 > 
 > > Also, in an automatic-follow mode in one of the views and after
 > > continuously typing only from some other view, the editor in the
 > > first view changed for me after 1.5 minutes !! That is huge time
 > > for a person to become an editor. I think some 10 seconds or so
 > > should be the time ?
 > 
 > This is probably because of the 'input.empty()' condition in the
 > polling function, on continuous typing tiles need to be updated and
 > so the editor update messages are not sent.
 > Do you have any suggestion that i could do here.

Ah, yes. I see now. You are right, it is because we want to prioritize tile rendering to all other things (mem. calculations, this thing etc.).

Anyhow, while looking the code to get rid of this somehow, how about if you add a method to IDocumentManager which all the child sessions will call whenever invalidate_cursor callback is called, and inside this method which will be in the Document class in Kit.cpp, it will manage some data structure there such that one can easily retrieve the child session with the highest speed. This would reduce the complexity to O(1) when retrieving the speed there and then we can safely move it outside of input.empty() block as it won't be impacting the performance much then.

Patch-set: 6
Label: Verified=0