Update patch set 1
Patch Set 1:

Hmm - so the 'rlimit' stuff is ancient, and is completely orthogonal to seccomp - I wonder why it is in the same file & namespace - it is clearly confusing. Could we split that into a new namespace at least (perhaps leave the code in-place). And we shouldn't have to disable that along with seccomp.

Wrt. seccomp - I'd be inclined just to make the function call a no-op on a seccomp machine instead - if possible (?)

IIRC if the seccomp-init fails it logs some scary looking message - we should certainly do that in this case too. "Warning this code was compiled without seccomp enabled, this setup is not recommended for production." or something =)

Patch-set: 1
Label: Code-Review=0