Update patch set 3
Patch Set 3:

I just noticed that protocol.txt is lagging far behind. Not only this, but other changes that you made to the protocol are missing from the protocol.txt as well. Like the 'adddoc' command in protocol.txt still shows the non-JSON format.

Can you update the documentation in your free time in a follow-up patch ?

Patch-set: 3