Update patch set 3
Patch Set 3:

I haven't tried it yet, but I can't see any conditional limiting the Ruler to only text documents, and not calc and impress. If it's missing please add it; we surely don't want it in calc and impress. A quick smoke test with calc and impress would be great to make sure we aren't breaking anything there.

Patch-set: 3
Label: Verified=0