Update patch set 5
Patch Set 5: Code-Review-1

Just tried it; looks impressive. But ...

Would be good to have units there on the ruler. Did you do some research around it ? How hard would it be to have units there ?

Other than that, do you think it would be possible to add support for first-line indent, paragraph indent and tab spaces (those three small triangles on desktop writer etc) in future ? Not that we are focusing on it right now, but surely we'd need it sometime in future. It'd be great if you could file bug for these small but important things separately in bugzilla alongwith any relevant information you've on them. (and please CC me).

I'm happy to merge this but no support in browsers other than chrome sounds like a blocker to me, and it'd be great if you could fix that so that I could merge this; atleast it should work in firefox also other than chrome (bonus points for making it work on IE and Edge as well). We can take care of other tasks (as mentioned above and as mentioned in your todo task list on gsoc proposal page) later

Patch-set: 5
Label: Code-Review=-1