Update patch set 5
Patch Set 5:

> Would be good to have units there on the ruler. Did you do some
 > research around it ? How hard would it be to have units there ?

yes, I did some research and it would be pretty easy. The current display is in cm but conversion to any other would be as easy as multiplication with appropriate factors.

Initially i was sending the length of the ruler, lmargin etc in the form of twips, and then using a complicated incorrect way to convert it to apt units and display it. 
But as of now, the values being sent in the callback are in units if 100th mm, which is much easier to convert and display.

 > Other than that, do you think it would be possible to add support
 > for first-line indent, paragraph indent and tab spaces (those three
 > small triangles on desktop writer etc) in future ? Not that we are
 > focusing on it right now, but surely we'd need it sometime in
 > future. It'd be great if you could file bug for these small but
 > important things separately in bugzilla alongwith any relevant
 > information you've on them. (and please CC me).

yes, it should be easy to do. I'll file bug reports for them.

 > I'm happy to merge this but no support in browsers other than
 > chrome sounds like a blocker to me, and it'd be great if you could
 > fix that so that I could merge this; atleast it should work in
 > firefox also other than chrome (bonus points for making it work on
 > IE and Edge as well). We can take care of other tasks (as mentioned
 > above and as mentioned in your todo task list on gsoc proposal
 > page) later

ok, i'll work on this ASAP.

Patch-set: 5
Label: Verified=0