Update patch set 2
Patch Set 2:

Looks mostly good. 

However, note that we'd need to input path of the suggested target, and not just the name of the file. And the way to get the path would be to ask the WOPI host. The way I can imagine it working is by delegating this UI dialog box to the WOPI host, which then will be responsible for the name of the file as well as path where the new file will be (since WOPI host UI will have all the idea about directory structure etc.). So we don't need to have any dialog UI in loleaflet. Just do something like how we do it for the revision history -- we send a post message to the outer frame and ask it to invoke the revision history UI. You will do exactly same, i.e., send a post message to the outer frame and ask it to invoke the save as dialog, that's it. The another post message API you would need is to receive the saveas command from the outer frame alongwith the path of the file (including the name of the file). Do check if there is such a thing already in the WOPI protocol. If yes, then reuse the protocol, otherwise, create your own post message names.

I think rest of the patch is fine/going in good direction.

Patch-set: 2