Update patch set 8
Patch Set 8:

Basically, what I wanted, as a last thing, is that we detect if the back-end supports ruler or not. If not, we do not add the ruler control to the map. This can be accomplished by rulerstate command values maybe, but I don't see a call to it at all.

Patch-set: 8
Label: Verified=0