Update patch set 5
Patch Set 5:

> Also, do you think it's possible to somehow get rid of repeated
 > units along the axes. Eg: 200 MB, 400 MB. Here MB is repeated; I
 > was thinking maybe if we could somehow do something different here
 > to avoid the redundancy here.

Taking inspiration from the conventional time series graph, I think it'd make sense to just put a heading the graph with units mentioned and then no need to mention the units on the Y-axis. Eg: For Network graph, we can have a heading like 'Memory consumed (in MB)' somewhere in the graph and then remove all the 'MB's on the Y-axis.

Patch-set: 5